// packages/charts/src/interfaces/enums.ts
export const Events = EventEnums
// ...
export enum AxisPositions {
	LEFT = 'left',
	RIGHT = 'right',
	TOP = 'top',
	BOTTOM = 'bottom'
}
// ...
export enum ScaleTypes {
	TIME = 'time',
	LINEAR = 'linear',
	LOG = 'log',
	LABELS = 'labels',
	LABELS_RATIO = 'labels-ratio'
}